[wip] sample bootstrapping code out of scaffolding
